pyproject.toml
[tool.poetry]
name = "ubic-test-task"
version = "0.1.0"
description = ""
authors = ["Nikita Y. <fmajesty@pm.me>"]

[tool.poetry.dependencies]
python = "^3.8"
lark-parser = "^0.11.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
